fix for AsyncVector Env
