+++ cyberwheel/utils/__init__.py
@@ -11,3 +11,4 @@ from cyberwheel.utils.train_cyberwheel import train_cyberwheel
 from cyberwheel.utils.run_cyberwheel import run_cyberwheel
 from cyberwheel.utils.run_visualization_server import run_visualization_server
 from cyberwheel.utils.set_seed import set_seed
+from cyberwheel.utils.async_call import async_call, make_env
